Revert moving replyWithHelp
