+++ src/slash-command-handler.ts
@@ -1,9 +1,8 @@
 'use strict';
 
 import { authLambdaEvent } from "./slack-auth";
-import { listContributions, LambdaResponse, postModalBlock } from "./shared/slack-interaction";
+import { listContributions, replyWithHelp, postModalBlock } from "./shared/slack-interaction";
 const { parse } = require('querystring');
-import { Config } from "./shared/config";
 
 interface SlackSlashCommandPayload {
     text: string;
@@ -50,55 +49,4 @@ export const handleSlashCommand = (event: any) => {
 
     return replyWithHelp();
 };
-
-/**
- * Replies with help text
- */
-const replyWithHelp = (): Promise<LambdaResponse> => {
-    // KLUDGE: environment should be mocked for tests, because Config is fail fast
-    let version;
-    let environment;
-
-    try {
-      version = Config.get('VERSION');
-      environment = Config.get('ENVIRONMENT');
-    } catch (e) {
-        console.error("Something went wrong with fetching config", e);
-    }
-
-    const helpMessage = [
-        "*Hi there!*",
-        "",
-        "My name is Ossi (a.k.a Ossitron-2000) :robot_face:, and I'm here to record your Open Source Contributions. :gem:",
-        "",
-        "First use slash command /ossi new and then fill the modal inputs. You can also type the description field after new-word",
-        "",
-        " * Description of the contribution",
-        " * URL of contribution",
-        " * Contribution month",
-        " * Contribution compensation from medium_200€/small_50€/no_compensation / competence_development_hours",
-        "",
-        "If you decide to submit, I will store the contribution to DynamoDB and notify my management channel for sanity check your contribution.",
-        "",
-        "When your contribution gets processed, I will notify you back.", ,
-        "",
-        "If you have questions about the process contact Valtteri Valovirta. If I'm broken contact Juho Friman.",
-        "",
-        "I have additional features under this slash command. My slash commands are all _ephemeral_ which means only you see the results. So feel free to shoot slash commands at any channel.",
-        "",
-        "`help` shows this help",
-        "`list` lists your submitted contributions",
-        "`new #description_here#`  subscribe new contribution",
-        "",
-        "I'm deployed into :aws-super-hero: AWS Cloud to `eu-north-1` region to Solita Sandbox account. I'm built of Node.js, Typescript, Serverless, Api Gateway, Lambda and DynamoDB.",
-        "",
-        "_Information about the policy_: https://intra.solita.fi/pages/viewpage.action?pageId=76514684",
-        "_My source code_: https://github.com/solita/ossi-bot",
-        `_Deployment_: ${version} ${environment}`
-    ].join('\n');
-    return Promise.resolve({
-        statusCode: 200,
-        body: helpMessage
-    })
-}
